3517 Refine AddCoordinateChannels transform (#3524)
* [DLMED] change to utility transforms


* [DLMED] enhance args
