+++ docs/source/transforms.rst
@@ -725,13 +725,6 @@ Spatial
     :members:
     :special-members: __call__
 
-`AddCoordinateChannels`
-"""""""""""""""""""""""
-.. autoclass:: AddCoordinateChannels
-    :members:
-    :special-members: __call__
-
-
 Smooth Field
 ^^^^^^^^^^^^
 
@@ -932,6 +925,12 @@ Utility
 `RandCuCIM`
 """""""""""
 .. autoclass:: RandCuCIM
+    :members:
+    :special-members: __call__
+
+`AddCoordinateChannels`
+"""""""""""""""""""""""
+.. autoclass:: AddCoordinateChannels
     :members:
     :special-members: __call__
 
@@ -1519,12 +1518,6 @@ Spatial (Dict)
     :members:
     :special-members: __call__
 
-`AddCoordinateChannelsd`
-""""""""""""""""""""""""
-.. autoclass:: AddCoordinateChannelsd
-    :members:
-    :special-members: __call__
-
 `GridDistortiond`
 """""""""""""""""
 .. image:: https://github.com/Project-MONAI/DocImages/raw/main/transforms/GridDistortiond.png
@@ -1764,6 +1757,12 @@ Utility (Dict)
 `RandCuCIMd`
 """"""""""""
 .. autoclass:: RandCuCIMd
+    :members:
+    :special-members: __call__
+
+`AddCoordinateChannelsd`
+""""""""""""""""""""""""
+.. autoclass:: AddCoordinateChannelsd
     :members:
     :special-members: __call__
