+++ monai/transforms/spatial/dictionary.py
@@ -29,7 +29,6 @@ from monai.networks.layers.simplelayers import GaussianFilter
 from monai.transforms.croppad.array import CenterSpatialCrop, SpatialPad
 from monai.transforms.inverse import InvertibleTransform
 from monai.transforms.spatial.array import (
-    AddCoordinateChannels,
     Affine,
     AffineGrid,
     Flip,
@@ -123,9 +122,6 @@ __all__ = [
     "ZoomDict",
     "RandZoomD",
     "RandZoomDict",
-    "AddCoordinateChannelsd",
-    "AddCoordinateChannelsD",
-    "AddCoordinateChannelsDict",
 ]
 
 GridSampleModeSequence = Union[Sequence[Union[GridSampleMode, str]], GridSampleMode, str]
@@ -1756,34 +1752,6 @@ class RandZoomd(RandomizableTransform, MapTransform, InvertibleTransform):
         return d
 
 
-class AddCoordinateChannelsd(MapTransform):
-    """
-    Dictionary-based wrapper of :py:class:`monai.transforms.AddCoordinateChannels`.
-    """
-
-    backend = AddCoordinateChannels.backend
-
-    def __init__(self, keys: KeysCollection, spatial_channels: Sequence[int], allow_missing_keys: bool = False) -> None:
-        """
-        Args:
-            keys: keys of the corresponding items to be transformed.
-                See also: :py:class:`monai.transforms.compose.MapTransform`
-            allow_missing_keys: don't raise exception if key is missing.
-            spatial_channels: the spatial dimensions that are to have their coordinates encoded in a channel and
-                appended to the input. E.g., `(1,2,3)` will append three channels to the input, encoding the
-                coordinates of the input's three spatial dimensions. It is assumed dimension 0 is the channel.
-
-        """
-        super().__init__(keys, allow_missing_keys)
-        self.add_coordinate_channels = AddCoordinateChannels(spatial_channels)
-
-    def __call__(self, data: Mapping[Hashable, NdarrayOrTensor]) -> Dict[Hashable, NdarrayOrTensor]:
-        d = dict(data)
-        for key in self.key_iterator(d):
-            d[key] = self.add_coordinate_channels(d[key])
-        return d
-
-
 class GridDistortiond(MapTransform):
     """
     Dictionary-based wrapper of :py:class:`monai.transforms.GridDistortion`.
@@ -1919,4 +1887,3 @@ RotateD = RotateDict = Rotated
 RandRotateD = RandRotateDict = RandRotated
 ZoomD = ZoomDict = Zoomd
 RandZoomD = RandZoomDict = RandZoomd
-AddCoordinateChannelsD = AddCoordinateChannelsDict = AddCoordinateChannelsd
