+++ tests/test_add_coordinate_channelsd.py
@@ -22,24 +22,20 @@ TESTS, TEST_CASES_ERROR_1, TEST_CASES_ERROR_2 = [], [], []
 for p in TEST_NDARRAYS:
     TESTS.append(
         [
-            {"spatial_channels": (1, 2, 3), "keys": ["img"]},
+            {"spatial_dims": (0, 1, 2), "keys": ["img"]},
             {"img": p(np.random.randint(0, 2, size=(1, 3, 3, 3)))},
             (4, 3, 3, 3),
         ]
     )
     TESTS.append(
-        [
-            {"spatial_channels": (1,), "keys": ["img"]},
-            {"img": p(np.random.randint(0, 2, size=(1, 3, 3, 3)))},
-            (2, 3, 3, 3),
-        ]
+        [{"spatial_dims": (0,), "keys": ["img"]}, {"img": p(np.random.randint(0, 2, size=(1, 3, 3, 3)))}, (2, 3, 3, 3)]
     )
 
     TEST_CASES_ERROR_1.append(
-        [{"spatial_channels": (3,), "keys": ["img"]}, {"img": p(np.random.randint(0, 2, size=(1, 3, 3)))}]
+        [{"spatial_dims": (2,), "keys": ["img"]}, {"img": p(np.random.randint(0, 2, size=(1, 3, 3)))}]
     )
     TEST_CASES_ERROR_2.append(
-        [{"spatial_channels": (0, 1, 2), "keys": ["img"]}, {"img": p(np.random.randint(0, 2, size=(1, 3, 3)))}]
+        [{"spatial_dims": (-1, 0, 1), "keys": ["img"]}, {"img": p(np.random.randint(0, 2, size=(1, 3, 3)))}]
     )
